refactor: templating을 담당하는 Templater 클래스 분리
